% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_calculator.R
\name{computeAIC}
\alias{computeAIC}
\title{Compute the approximate AIC}
\usage{
computeAIC(model)
}
\arguments{
\item{model}{the \code{mixedMemModel} object for which the AIC will be calculated.}
}
\value{
\code{computeBIC} returns the approximate AIC value, a real number.
}
\description{
\code{computeAIC} computes the approximate AIC of a given \code{mixedMemModel}, where the lower bound on the log-likelihood
(also called ELBO) is used instead of the intractable true log-likelihood.
}
\details{
\eqn{AIC = -2 ELBO + 2 * p}

where p is the number of estimated parameters.
}

