% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropyPlot.R
\name{entropyPlot}
\alias{entropyPlot}
\title{Mixed Membership Visualization}
\usage{
entropyPlot(model, col = "white", main = "Exponentiated Entropy",
  xlab = "Effective Profiles")
}
\arguments{
\item{model}{the \code{mixedMemModel} the model to be plotted}

\item{col}{the color of the bars}

\item{main}{the main title}

\item{xlab}{the label for the x-axis}
}
\description{
\code{entropyPlot} plots a histogram of the entropy of each individuals estimated mixed membership. The entropy is
is roughly the number of profiles needed to model each individual (Gormley and Murphy 2014; White et al 2012).
}

