% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPosteriorEstimates.R
\name{getPosteriorEstimates}
\alias{getPosteriorEstimates}
\title{Mixed Membership Post-Processing for MCMC method
   
\code{getPosteriorEstimates}}
\usage{
getPosteriorEstimates(model, fileNames, whichWrite = c(1, 1, 1, 0, 0, 1, 0))
}
\arguments{
\item{model}{the fitted \code{mixedMemModelMCMC} object.}

\item{fileNames}{a vector of strings containing the file paths for the MCMC output
The string should be of length 7 for theta, alpha, ksi, lambda, Z, P and StayerStatus.
If the particular parameter was not recorded, insert an empty string ("") as a placeholder}

\item{whichWrite, }{a vector of length 7 indicating which files to be read. 1 indicates read,
0 indicates skip}
}
\description{
Mixed Membership Post-Processing for MCMC method
   
\code{getPosteriorEstimates}
}

