% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{gmv_theta}
\alias{gmv_theta}
\title{Point estimates from Gross and Manrique-Vallier 2014}
\usage{
data(gmv_theta)
}
\description{
\code{gmv_theta} contains the point estimates for the sub-populaton response probabilities, \eqn{\theta} from the original Gross and
Manrique-Vallier 2014 analysis. It is a 3 by 19 by 3 dimensional array with
the first dimension corresponding to each latent sub-population, the second dimension corresponding to
each variable, and the 3rd dimension corresponding to response choices (Agree, Can't Decide,
Disagree). Gross and Manrique-Vallier do not report results for group 3 for reasons discussed in the vignette
so all values reported for group 3 in theta[1, , ] are 0.
}
\references{
Gross, J.H.; Manrique-Vallier, D. "A Mixed-Membership Approach to the Assessment of Political Ideology from Survey Responses."
in Airoldi, E. M., Blei, D. M., Erosheva, E. A., and Fienberg, S. E.
Handbook of Mixed Membership Models and Its Applications. Chapman & Hall/CRC, 2014
}

