% Generated by roxygen2 (4.0.2): do not edit by hand
\name{heldOut}
\alias{heldOut}
\title{Calculate ELBO from Held out Data}
\usage{
heldOut(model)
}
\arguments{
\item{model}{a \code{mixedMemModel} object created by the \code{mixedMemModel} constructor
which has been fit using mmVarFit}
}
\description{
Theta and alpha are fixed while estimating the group memberships (phi) and context (delta)
for a set of held out data. Used in cross validation
}

