% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellingerDistances.R
\name{hellingerDistances}
\alias{hellingerDistances}
\title{Mixed membership models post-processing}
\usage{
hellingerDistances(model)
}
\arguments{
\item{model}{the \code{mixedMemModelVI} object that will be plotted.}
}
\description{
\code{hellingerDistances} computes hellinger distances between each sub-population for each variable
}

