% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmMCMCFit.R
\name{mmMCMCFit}
\alias{mmMCMCFit}
\title{Fit Mixed Membership models using a Metropolis Hastings within Gibbs MCMC sampler}
\usage{
mmMCMCFit(model, burnIn = 20000, samples = 1000, thin = 10, print = 100,
  fileNames, newFiles = 1, omega = 100, eta = 1, whichWrite = c(1, 1, 1,
  0, 0, 1, 0))
}
\arguments{
\item{model}{a \code{mixedMemModelMCMC} object created by the \code{mixedMemModelMCMC} constructor}

\item{burnIn}{non-negative integer indicating the number of burn in samples before recording the first sample}

\item{samples}{positive integer indicating the number of samples to record}

\item{thin}{positive integer indicating how to thin the samples}

\item{print}{positive integer indicating how often to print an update to the R console}

\item{fileNames}{list of files names to write samples}

\item{newFiles}{0 if samples should be appended to existing files; 1 if samples should overwrite any existing files}

\item{omega}{tuning parameter for MH step for alpha.}

\item{eta}{tuning parameter for MH step for ksi}

\item{whichWrite}{which parameters to write to disk. Writing the individual parameters can significantly increase computational
time when the number of individuals is large}

\item{extended}{boolean whether to estimate the extended model or not}
}
\description{
Takes samples form the posterior distribution of a mixed membership model using a 
Metropolis-Hastings within Gibbs sampler.
}
\details{
\code{mmMCMCFit} draws samples from the posterior distribution of a mixed membership model
given a set of observed data. The sampler takes Metropolis Hastings steps to sample the \eqn{\alpha_0}
and \eqn{\xi} parameters and samples the remaining \eqn{\theta}, \eqn{\lambda_i} and \eqn{Z} parameters
with a Gibbs sampler.
}
\seealso{
mixedMemModelMCMC
}

