% Generated by roxygen2 (4.0.2): do not edit by hand
\name{permuteLabels}
\alias{permuteLabels}
\title{Mixed Membership Post-Processing}
\usage{
permuteLabels(model, perm)
}
\arguments{
\item{model}{the fitted \code{mixedMemModel} object}

\item{perm}{the permutation by which to relabel the \code{mixedMemModel} object. Must be a vector of length model$K.}
}
\value{
a permuted \code{mixedMemModel} object
}
\description{
Mixed Membership models are invariant to permutations of the sub-population labels and the ordering of the labels
is dependent on the initialization of the variational EM algorithim.
Given a permutation, which can be found using \code{findLabels},
the method returns a \code{mixedMemModel} object with permuted labels
}
\seealso{
findLabels
}

