% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPosteriorEstimates.R
\name{plotPosterior}
\alias{plotPosterior}
\title{Mixed Membership Post-Processing for MCMC method
   
\code{plotPosterior}}
\usage{
plotPosterior(model, fileName, parameter, new = T, type = "trace", j = 1,
  k = 1, v = 1, s = 1, ...)
}
\arguments{
\item{model}{the fitted \code{mixedMemModelMCMC} object.}

\item{fileName}{a string containing the file paths for the MCMC output}

\item{parameter}{which parameter to plot}

\item{new}{whether to plot on existing plot or new plot}

\item{type}{histogram or trace plot}

\item{...}{additional commands passed to plotting function}
}
\description{
Mixed Membership Post-Processing for MCMC method
   
\code{plotPosterior}
}

