% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedMemModel.R
\name{summary.mixedMemModelMCMC}
\alias{summary.mixedMemModelMCMC}
\title{Summary of a Mixed Membership Model}
\usage{
\method{summary}{mixedMemModelMCMC}(object, ...)
}
\arguments{
\item{object}{the \code{mixedMemModelMCMC} object to be summarized}

\item{...}{additional parameters}
}
\description{
Generic S3 summary function for \code{mixedMemModelMCMC} class.
}
\details{
\code{summary} provides a summary of the given \code{mixedMemModelMCMC} object. The function 
prints the ELBO, the dimensions of the model and each variable type.
}
\seealso{
mixedMemModelMCMC
}

