% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedMemModel.R
\name{summary.mixedMemModelVarInf}
\alias{summary.mixedMemModelVarInf}
\title{Summary of a Mixed Membership Model}
\usage{
\method{summary}{mixedMemModelVarInf}(object, ...)
}
\arguments{
\item{object}{the \code{mixedMemModelVarInf} object to be summarized}

\item{...}{additional parameters}
}
\description{
Generic S3 summary function for \code{mixedMemModelVarInf} class.
}
\details{
\code{summary} provides a summary of the given \code{mixedMemModelVarInf} object. The function 
prints the ELBO, the dimensions of the model and each variable type.
}
\seealso{
mixedMemModelVI
}

