% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedMemModel.R
\name{summary.mixedMemModelVI}
\alias{summary.mixedMemModelVI}
\title{Summary of a Mixed Membership Model}
\usage{
\method{summary}{mixedMemModelVI}(object, ...)
}
\arguments{
\item{object}{the \code{mixedMemModelVI} object to be summarized}

\item{...}{additional parameters}
}
\description{
Generic S3 summary function for \code{mixedMemModelVI} class.
}
\details{
\code{summary} provides a summary of the given \code{mixedMemModelVI} object. The function 
prints the ELBO, the dimensions of the model and each variable type.
}
\seealso{
mixedMemModelVI
}

