% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedMemModel.R
\name{theta.table}
\alias{theta.table}
\title{Print the theta estimates for each sub-population}
\usage{
theta.table(model, digits = 3, top.level = "group")
}
\arguments{
\item{model}{the \code{mixedMemModel} who's theta will be printed.}
}
\description{
Print the theta estimates for each sub-population to the R console in a more
aesthetically pleasing manner
}
\seealso{
mixedMemModel, vizTheta
}

