% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{gmv_theta}
\alias{gmv_theta}
\title{Point estimates from Gross and Manrique-Vallier 2014}
\usage{
data(gmv_theta)
}
\description{
The point estimates for theta from the original Gross and
Manrique-Vallier 2014 analysis. It is a 3 by 19 by 3 dimensional array with
the first dimension corresponding to group, the second dimension corresponding to
variable, and the 3rd dimension corresponding to candidate choices (Agree, Can't Decide,
Disagree).
}
\references{
Gross, J.H. and Manrique-Vallier, D. "A Mixed-Membership Approach to the Assessment of Political Ideology from Survey Responses."
in Airoldi, E. M., Blei, D. M., Erosheva, E. A., and Fienberg, S. E.
Handbook of Mixed Membership Models and Its Applications. Chapman & Hall/CRC, 2014
}

