% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.mixedMemModel}
\alias{plot.mixedMemModel}
\title{Plot a Mixed Membership Model}
\usage{
\method{plot}{mixedMemModel}(x, ...)
}
\arguments{
\item{x}{the mixedMemModel object to be plotted}

\item{...}{additional parameters passed to vizTheta}
}
\description{
Visual representation of a mixedMemModel object
}
\details{
Calls the vizTheta function to plot the values of \eqn{\theta} for each variable and sub-population.
}
\seealso{
vizTheta, mixedMemModel
}

