% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPosteriorEstimates.R
\name{plotPosterior}
\alias{plotPosterior}
\title{Mixed Membership Post-Processing for MCMC method
   
\code{plotPosterior}}
\usage{
plotPosterior(model, fileName, parameter, addToExisting = F,
  plotType = "trace", j = 1, k = 1, v = 1, s = 1, ...)
}
\arguments{
\item{model}{the fitted \code{mixedMemModelMCMC} object.}

\item{fileName}{a string containing the file paths for the MCMC output}

\item{parameter}{which parameter to plot}

\item{addToExisting}{whether to plot on existing plot or new plot}

\item{j}{the jth element to plot}

\item{k}{the kth element to plot}

\item{v}{the vth element to plot}

\item{s}{the sth element to plot}

\item{...}{additional commands passed to plotting function}

\item{type}{histogram or trace plot}
}
\description{
Mixed Membership Post-Processing for MCMC method
   
\code{plotPosterior}
}

