% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.mixedMemModel}
\alias{summary.mixedMemModel}
\title{Summary of a Mixed Membership Model}
\usage{
\method{summary}{mixedMemModel}(object, ...)
}
\arguments{
\item{object}{the mixedMemModel object to be summarized}
}
\description{
Summary of a mixedMemModel object
}
\details{
Prints summary data for a mixedMemModel object which includes the
ELBO, the dimensions of the model and details about each variable.
}
\seealso{
mixedMemModel
}

