% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vizTheta}
\alias{vizTheta}
\title{Mixed Membership Visualization}
\usage{
vizTheta(model, truth = NULL, main = "Estimated Theta", highlight = NULL,
  varNames = NULL, groupNames = NULL)
}
\arguments{
\item{model}{the \code{mixedMemModel} object that will be plotted}

\item{truth}{optional argument if ground truth is available}

\item{main}{optional title argument}

\item{highlight}{matrix with 2 columns specifying any specific variable/sub-population combinations
to be highlighted in the plot}

\item{varNames}{vector of length J specifying labels for each variable}

\item{groupNames}{vector of length K specifying labels for each sub-population}
}
\description{
Plots \eqn{\theta}, the parameters which govern the distributions of variables in a mixed membership model. \eqn{\theta_{j,k}} governs the
distribution of variable j for sub-population k. Estimated distributions are shown with black bars. The ground truth theta (if available)
is shown in green dots.
}

